Rename runtimes folder to gameTypes
