+++ src/app/features/gameTypes/versus/pixi/PixiRenderer.tsx
