+++ src/app/features/gameTypes/versus/pixi/ResponsiveStage.tsx
